Refactor author into its own struct
This commit removes a previous TODO about returning a tuple during
parsing. It does so my refactoring the `name` and `email` fields of a
`Commit` into an `Author` struct.

Although it's a straightforward change, it hinted at a possibly better
pattern. My next changes were going to involve the `newtype` idiom --
honestly one of my favorite for improving day-to-day coding. If every
field of a `Commit` ends up with its own type, why not couple parsing
at __that__ granularity?

For example, the number of files, insertions, and deletions that occur
for a commit are all fields. Each already has a regular expression
defined. Why not add the regular expression, or parsing trait, of each
field at its definition? Nestability should follow.